Add assert statements to getData() and getTaskData() methods
The program will run without errors
even if the output of CustomDateTime.getData() and Task.getTaskData() could not correctly initialize the same task.

This could cause saved data to be
corrupted without the user noticing.

Adding assert statements allow us to verify
the correctness of these methods and notify
the developer if any issue occurs.

Assert statements are used instead of
exceptions as the method's output should
be correct regardless of user input.