Add recurring tasks for deadline and event tasks
Many tasks occur periodically and users have to
add multiple similar tasks to the list to track them.

This is tedious for the user and results in a
messy and cluttered task list.

Users now have the ability to set recurring
tasks which replace themselves with a
similar task a set number of days later once it
has been marked as done.

This allows users to track the next instance of
the task at all times without filling the task
list with multiple copies of similar tasks.